refactor: transfer HSL definition to its own package
